reworking memory reading
changed signature of methods to read memory in different addressing
modes. When reading the memory, the addresses have actually be read from
memory. Therefore I changed the naming from `get_addr...` to
`read_mem...` and read the memory directly.

I suspect there is a danger of having to refactor that again if I need
more than two operands for an instruction. Not sure yet how this is
handled.